new doc
All sections completed except Methods for Telomere Resetting.

My next tasks are to
Assume the goal is to use the cycloastragenol / ta-65 compound and put it on the f13 protein (molecular docking) and make a new compound that is similar to ta-65 but is made out of simpler molecules, easier to chemically synthesize, this is simple I think. 
and
most importantly, the ADME properties are incredible to the point that just taking this compound would be enough for biological immortality (the method is by "sticking" to the protein for an extended period, because we already know that a relatively weak telomerase increaser like ta-65 can still increase telomere length.).
so we need machine learning tools to predict the ADME properties and improve them.
Would this be safe? Well we are adapting from the ta-65 molecule, which has lots of safety data, and is thus very useful for in silico analysis and machine learning.